# docs/api-docs.json
{"openapi":"3.0.1","info":{"title":"Cinema API","description":"API for managing cinemas.","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/schedules/{id}":{"get":{"tags":["schedule-controller"],"summary":"Get schedule by id","description":"Returns a single schedule, if one with the given id exists.","operationId":"getSchedulesById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScheduleDTO"}}}}}},"put":{"tags":["schedule-controller"],"summary":"Update a schedule","description":"Updates a schedule.","operationId":"updateSchedule","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScheduleDTO"}}}}}},"delete":{"tags":["schedule-controller"],"summary":"Delete a schedule","description":"Deletes a schedule.","operationId":"deleteSchedule","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prices/seats/{name}":{"get":{"tags":["price-controller"],"summary":"Get seat price by name","description":"Returns a single seat price, if one with the given name exists.","operationId":"getSeatPriceByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SeatPrice"}}}}}},"put":{"tags":["price-controller"],"summary":"Update a seat price","description":"Updates a seat price.","operationId":"updateSeatPrice","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeatPrice"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SeatPrice"}}}}}}},"/api/prices/reservations/{name}":{"get":{"tags":["price-controller"],"summary":"Get reservation price by name","description":"Returns a single reservation price, if one with the given name exists.","operationId":"getReservationPriceByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationPrice"}}}}}},"put":{"tags":["price-controller"],"summary":"Update a reservation price","description":"Updates a reservation price.","operationId":"updateReservationPrice","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationPrice"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationPrice"}}}}}}},"/api/prices/movies/{name}":{"get":{"tags":["price-controller"],"summary":"Get movie price by name","description":"Returns a single movie price, if one with the given name exists.","operationId":"getMoviePriceByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MoviePrice"}}}}}},"put":{"tags":["price-controller"],"summary":"Update a movie price","description":"Updates a movie price.","operationId":"updateMoviePrice","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoviePrice"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MoviePrice"}}}}}}},"/api/movies/{id}":{"get":{"tags":["movie-controller"],"summary":"Get movie by id","description":"Returns a single movie, if one with the given id exists.","operationId":"findMovieById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovieDTO"}}}}}},"put":{"tags":["movie-controller"],"summary":"Update a movie","description":"Updates a movie.","operationId":"updateMovie","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovieDTO"}}}}}},"delete":{"tags":["movie-controller"],"summary":"Delete a movie","description":"Soft deletes a movie, by flagging it as inactive.","operationId":"deleteMovie","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/halls/{id}":{"get":{"tags":["hall-controller"],"summary":"Get hall by id","description":"Returns a single hall, if one with the given id exists.","operationId":"getHallById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HallDTO"}}}}}},"put":{"tags":["hall-controller"],"summary":"Update a hall","description":"Updates a hall.","operationId":"updateHall","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HallDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HallDTO"}}}}}},"delete":{"tags":["hall-controller"],"summary":"Delete a hall","description":"Soft deletes a hall, by flagging it as inactive.","operationId":"deleteHall","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HallDTO"}}}}}}},"/api/cinemas/{id}":{"get":{"tags":["cinema-controller"],"summary":"Get cinema by id","description":"Returns a single cinema, if one with the given id exists.","operationId":"findCinemaById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CinemaDTO"}}}}}},"put":{"tags":["cinema-controller"],"summary":"Update a cinema","description":"Updates a cinema.","operationId":"updateCinema","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CinemaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CinemaDTO"}}}}}},"delete":{"tags":["cinema-controller"],"summary":"Delete a cinema","description":"Soft deletes a cinema, by flagging it as inactive.","operationId":"deleteCinema","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CinemaDTO"}}}}}}},"/api/user-with-role":{"post":{"tags":["user-with-role-controller"],"summary":"Add a new UserWithRoles user","description":"If a default role is defined (app.default-role ), this role will be assigned to the user.","operationId":"addUserWithRoles","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithRolesRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}}}}},"/api/schedules":{"get":{"tags":["schedule-controller"],"summary":"Get all schedules","description":"Returns a list of all schedules.","operationId":"getSchedules","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScheduleDTO"}}}}}}},"post":{"tags":["schedule-controller"],"summary":"Create a schedule","description":"Creates a new schedule.","operationId":"createSchedule","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScheduleDTO"}}}}}}},"/api/reservations":{"post":{"tags":["reservation-controller"],"summary":"Create a reservation","description":"Creates a new reservation.","operationId":"createReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationReqDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResDTO"}}}}}}},"/api/reservations/prices":{"post":{"tags":["reservation-controller"],"summary":"Calculate price","description":"Calculates the price for a reservation based on the provided reservation request.","operationId":"calculatePrice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationReqDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationPriceCalcDTO"}}}}}}},"/api/movies":{"get":{"tags":["movie-controller"],"summary":"Get all movies","description":"Returns a list of all movies.","operationId":"getAllMovies","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieDTO"}}}}}}},"post":{"tags":["movie-controller"],"summary":"Create a movie","description":"Creates a new movie","operationId":"addMovie","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovieDTO"}}}}}}},"/api/halls":{"get":{"tags":["hall-controller"],"summary":"Get all halls","description":"Returns a list of all halls.","operationId":"getHalls","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HallDTO"}}}}}}},"post":{"tags":["hall-controller"],"summary":"Create a hall","description":"Creates a new hall.","operationId":"addHall","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HallDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HallDTO"}}}}}}},"/api/cinemas":{"get":{"tags":["cinema-controller"],"summary":"Get all cinemas","description":"Returns a list of all cinemas.","operationId":"getAllCinemas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CinemaDTO"}}}}}}},"post":{"tags":["cinema-controller"],"summary":"Create a cinema","description":"Creates a new cinema.","operationId":"addCinema","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CinemaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CinemaDTO"}}}}}}},"/api/auth/logout":{"post":{"tags":["authentication-controller"],"summary":"Logout","description":"Used to invalidate token on logout","operationId":"logout","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/login":{"post":{"tags":["authentication-controller"],"summary":"Login","description":"Use this to login and get a token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/user-with-role/remove-role/{username}/{role}":{"patch":{"tags":["user-with-role-controller"],"summary":"Removes a role from a UserWithRoles","description":"Caller must be authenticated with the role ADMIN","operationId":"removeRole","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}},{"name":"role","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}}}}},"/api/user-with-role/add-role/{username}/{role}":{"patch":{"tags":["user-with-role-controller"],"summary":"Add a role to a UserWithRoles","description":"Caller must be authenticated with the role ADMIN","operationId":"addRole","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}},{"name":"role","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}}}}},"/api/schedules/{date}/movies/{movieId}/cinemas/{cinemasId}":{"get":{"tags":["schedule-controller"],"summary":"Get schedules by date, movie and cinema","description":"Returns a list of schedules for a specific date, movie id and cinema id.","operationId":"getSchedulesByDateAndCinemaIdAndMovieId","parameters":[{"name":"date","in":"path","required":true,"schema":{"type":"string","format":"date"}},{"name":"movieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"cinemasId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScheduleDTO"}}}}}}}},"/api/reservations/schedules/{id}/seats":{"get":{"tags":["reservation-controller"],"summary":"Get reservation by schedule id","description":"Returns a list of reservations for a schedule with the given id.","operationId":"findAllReservedSeatsByScheduleId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeatDTO"}}}}}}}},"/api/prices/seats":{"get":{"tags":["price-controller"],"summary":"Get all seat prices","description":"Returns a list of all seat prices.","operationId":"getAllSeatPrices","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeatPrice"}}}}}}}},"/api/prices/reservations":{"get":{"tags":["price-controller"],"summary":"Get all reservation prices","description":"Returns a list of all reservation prices.","operationId":"getAllReservationPrices","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservationPrice"}}}}}}}},"/api/prices/movies":{"get":{"tags":["price-controller"],"summary":"Get all movie prices","description":"Returns a list of all movie prices.","operationId":"getAllMoviePrices","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MoviePrice"}}}}}}}},"/api/halls/{hallName}/cinemas/{cinemaName}":{"get":{"tags":["hall-controller"],"summary":"Get hall by name and cinema name","description":"Returns a single hall, if one with the given name and cinema name exists.","operationId":"findByNameAndCinemaName","parameters":[{"name":"hallName","in":"path","required":true,"schema":{"type":"string"}},{"name":"cinemaName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HallDTO"}}}}}}},"/api/halls/cinema/{cinemaId}":{"get":{"tags":["hall-controller"],"summary":"Get halls by cinema id","description":"Returns a list of halls in a cinema with the given id.","operationId":"findHallsByCinemaId","parameters":[{"name":"cinemaId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HallDTO"}}}}}}}},"/api/cinemas/movie/{movieId}":{"get":{"tags":["cinema-controller"],"summary":"Get cinemas by movie id","description":"Returns a list of cinemas showing a movie with the given id.","operationId":"findCinemasByMovieId","parameters":[{"name":"movieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CinemaDTO"}}}}}}}}},"components":{"schemas":{"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}}},"ScheduleDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"startTime":{"$ref":"#/components/schemas/LocalTime"},"endTime":{"$ref":"#/components/schemas/LocalTime"},"date":{"type":"string","format":"date"},"movieTitle":{"type":"string"},"hallName":{"type":"string"},"cinemaName":{"type":"string"},"is3d":{"type":"boolean"},"isLongMovie":{"type":"boolean"}}},"SeatPrice":{"type":"object","properties":{"name":{"type":"string"},"amount":{"type":"number","format":"double"},"unit":{"type":"string"}}},"ReservationPrice":{"type":"object","properties":{"name":{"type":"string"},"amount":{"type":"number","format":"double"},"unit":{"type":"string"}}},"MoviePrice":{"type":"object","properties":{"name":{"type":"string"},"amount":{"type":"number","format":"double"},"unit":{"type":"string"}}},"MovieDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"description":{"type":"string"},"releaseDate":{"type":"string","format":"date"},"duration":{"type":"integer","format":"int32"},"imageUrl":{"type":"string"},"language":{"type":"string"},"genre":{"type":"string"},"director":{"type":"string"},"cast":{"type":"string"},"active":{"type":"boolean"}}},"HallDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"noOfRows":{"type":"integer","format":"int32"},"noOfColumns":{"type":"integer","format":"int32"},"imageUrl":{"type":"string"},"cinemaId":{"type":"integer","format":"int32"},"active":{"type":"boolean"}}},"CinemaDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"city":{"type":"string"},"street":{"type":"string"},"description":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"imageUrl":{"type":"string"},"halls":{"type":"array","items":{"$ref":"#/components/schemas/HallDTO"}},"active":{"type":"boolean"}}},"UserWithRolesRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"}}},"UserWithRolesResponse":{"type":"object","properties":{"userName":{"type":"string"},"roleNames":{"type":"array","items":{"type":"string"}},"email":{"type":"string"}}},"ReservationReqDTO":{"type":"object","properties":{"scheduleId":{"type":"integer","format":"int32"},"seatIndexes":{"uniqueItems":true,"type":"array","items":{"type":"integer","format":"int32"}}}},"ReservationResDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"scheduleId":{"type":"integer","format":"int32"},"tickets":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/TicketDTO"}},"reservationDate":{"type":"string","format":"date"},"feeOrDiscount":{"type":"number","format":"double"}}},"TicketDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"seatId":{"type":"integer","format":"int32"},"price":{"type":"number","format":"double"},"reservationId":{"type":"integer","format":"int32"}}},"ReservationPriceCalcDTO":{"type":"object","properties":{"scheduleId":{"type":"integer","format":"int32"},"tickets":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/TicketPriceCalcDTO"}},"feeOrDiscount":{"type":"number","format":"double"}}},"TicketPriceCalcDTO":{"type":"object","properties":{"seatId":{"type":"integer","format":"int32"},"seatIndex":{"type":"integer","format":"int32"},"price":{"type":"number","format":"double"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"username":{"type":"string"},"token":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}},"SeatDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"seatIndex":{"type":"integer","format":"int32"},"seatPriceName":{"type":"string"},"hallId":{"type":"integer","format":"int32"}}}}}}